reorganizing code and tests to extract ANTLR deps from core
