+++ antlr/build.gradle
@@ -20,6 +20,10 @@ dependencies {
     antlr "org.antlr:antlr4:$antlr_version"
     api project(":core")
     implementation "org.antlr:antlr4-runtime:$antlr_version"
+    implementation "org.jetbrains.kotlin:kotlin-reflect:$kotlin_version"
+    implementation "org.jetbrains.kotlin:kotlin-test:$kotlin_version"
+
+    testImplementation "org.jetbrains.kotlin:kotlin-test-junit:$kotlin_version"
 }
 
 generateTestGrammarSource {
