+++ core/src/main/kotlin/com/strumenta/kolasu/language/LanguageModule.kt
@@ -2,7 +2,7 @@ package com.strumenta.kolasu.language
 
 import com.strumenta.kolasu.model.Node
 import com.strumenta.kolasu.model.walkChildren
-import com.strumenta.kolasu.parsing.KolasuParser
+import com.strumenta.kolasu.parsing.ASTParser
 
 // TODO Alessio this is a work in progress
 class Unparser {
@@ -28,4 +28,4 @@ class Unparser {
     }
 }
 
-class LanguageModule(val parser: KolasuParser<*, *, *>, val unparser: Unparser)
+class LanguageModule(val parser: ASTParser<*>, val unparser: Unparser)
