+++ core/src/test/kotlin/com/strumenta/kolasu/model/PositionTest.kt
@@ -1,19 +1,10 @@
 package com.strumenta.kolasu.model
 
-import com.strumenta.kolasu.parsing.ParseTreeOrigin
-import com.strumenta.kolasu.parsing.toPosition
-import com.strumenta.simplelang.SimpleLangLexer
-import com.strumenta.simplelang.SimpleLangParser
-import org.antlr.v4.runtime.CharStreams
-import org.antlr.v4.runtime.CommonTokenStream
-import org.antlr.v4.runtime.tree.ParseTree
 import kotlin.test.assertEquals
 import kotlin.test.assertFails
 import kotlin.test.assertFalse
 import kotlin.test.assertTrue
 import org.junit.Test as test
-
-data class MySetStatement(val specifiedPosition: Position? = null) : Node(specifiedPosition)
 
 class PositionTest {
 
@@ -148,29 +139,6 @@ class PositionTest {
         assertFalse("contains should return false with node after") { position.contains(after) }
     }
 
-    @test
-    fun parserRuleContextPosition() {
-        val code = "set foo = 123"
-        val lexer = SimpleLangLexer(CharStreams.fromString(code))
-        val parser = SimpleLangParser(CommonTokenStream(lexer))
-        val cu = parser.compilationUnit()
-        val setStmt = cu.statement(0) as SimpleLangParser.SetStmtContext
-        val pos = setStmt.toPosition()
-        assertEquals(Position(Point(1, 0), Point(1, 13)), pos)
-    }
-
-    @test
-    fun positionDerivedFromParseTreeNode() {
-        val code = "set foo = 123"
-        val lexer = SimpleLangLexer(CharStreams.fromString(code))
-        val parser = SimpleLangParser(CommonTokenStream(lexer))
-        val cu = parser.compilationUnit()
-        val setStmt = cu.statement(0) as SimpleLangParser.SetStmtContext
-        val mySetStatement = MySetStatement()
-        mySetStatement.origin = ParseTreeOrigin(setStmt)
-        assertEquals(Position(Point(1, 0), Point(1, 13)), mySetStatement.position)
-    }
-
     @test fun illegalPositionAccepted() {
         Position(Point(10, 1), Point(5, 2), validate = false)
     }
@@ -179,17 +147,4 @@ class PositionTest {
     fun illegalPositionNotAccepted() {
         Position(Point(10, 1), Point(5, 2), validate = true)
     }
-
-    @test fun parserTreePosition() {
-        val code = "set foo = 123"
-        val lexer = SimpleLangLexer(CharStreams.fromString(code))
-        val parser = SimpleLangParser(CommonTokenStream(lexer))
-        val cu: ParseTree = parser.compilationUnit()
-        val pos = cu.toPosition()
-        assertEquals(Position(Point(1, 0), Point(1, 13)), pos)
-    }
-
-//    @test(expected = Exception::class) fun illegalPositionNotAcceptedByDefault() {
-//        val p = Position(Point(10, 1), Point(5, 2))
-//    }
 }
