+++ playground/src/main/kotlin/com/strumenta/kolasu/playground/PlaygroundExampleGenerator.kt
@@ -68,9 +68,9 @@ fun ParsingResult<*>.saveForPlayground(
         jsonObject.addProperty("name", name)
         jsonObject.addProperty("code", code)
         jsonObject.add("ast", ast)
-        if (firstStage != null) {
-            jsonObject.addProperty("parsingTime", firstStage!!.time)
-        }
+//        if (firstStage != null) {
+//            jsonObject.addProperty("parsingTime", firstStage!!.time)
+//        }
         jsonObject.addProperty("astBuildingTime", time)
         writer.write(jsonObject.toString())
     } finally {
